src/requirements.txt
fastapi
deta
pyjwt
